5198: Barrel files refactor: meta/expressions
