+++ src/meta/expressionEvaluator/index.ts
@@ -1,1 +1,2 @@
 export { ExpressionEvaluator } from './expressionEvaluator'
+export { Member } from 'meta/expressionEvaluator/member'
