--- src/meta/expressions/index.ts
@@ -1,1 +1,0 @@
-export { Member } from './member'
